Add OpenDNSFilter
The OpenDNSFilter will add a 'osxcollector_opendns' to lines that have a suspicious domain in the 'osxcollector_domains' key.
Running the OpenDNS filter requires having already run the DomainsFilter.

This also introduces a concept of a config file for filters. Each filter can have its own config in a YAML section with the
same name as a the class. This allows for stuff like API keys, whitelists, etc.